The Code of Music is LIVE!
This contains a working algorithm for the naïve music generator. It
includes 4 simple pentatonic scales that can change the nature of the
improvisation. More might be added in the future.

In addition, thought was put into including indico’s sentiment
recognition into choosing the scale, but due to lack of time, it was
purponde to a later time.